test: incorporate reviewr commnent
separate out parameter update checking function into
seperate module.
