+++ tests/ludwig/modules/test_tabnet_modules.py
@@ -5,7 +5,7 @@ import torch
 
 from ludwig.modules.tabnet_modules import AttentiveTransformer, FeatureBlock, FeatureTransformer, TabNet
 from ludwig.utils.entmax import sparsemax
-from tests.integration_tests.utils import check_module_parameters_updated
+from tests.integration_tests.parameter_update_utils import check_module_parameters_updated
 
 RANDOM_SEED = 67
